started support for snapshots

+++ src/NStore/Aggregates/Changeset.cs
@@ -6,7 +6,7 @@ namespace NStore.Aggregates
     public sealed class Changeset : IHeadersAccessor
     {
         public Object[] Events { get; private set; }
-        public long Version { get; private set; }
+        public int Version { get; private set; }
         public IDictionary<string, object> Headers { get; private set; }
         public bool IsEmpty => Events.Length == 0;
 
@@ -15,7 +15,7 @@ namespace NStore.Aggregates
             Headers = new Dictionary<string, object>();
         }
 
-        public Changeset(long version, params object[] events) : this()
+        public Changeset(int version, params object[] events) : this()
         {
             this.Version = version;
             this.Events = events;
